requirements.txt
opencv-python>=4.9.0
mediapipe>=0.10.14
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
pyttsx3>=2.90
